New view created
Created Alerts, Constraints, and a new view
This new view will be called from the Main and shows content about the project
Changes in the "MainController" where it was included a new method to call the new view, it will change save the value from the initial view and after add the initial value and the new value.
Save the old value is important to be able to back to the beginning.
This method will be synchronized to grant the process won't be stopped during the execution